requirements.txt
jass_kit
flask
numpy
scikit-learn
joblib
